Fix rendering bug in xml_transform.py
ElementTree will print an empty element as self-closing, as in
'<span id="bar" />'. But when the element is rendered on the page, it
is treated as an open tag that wraps all of the sibling elements that
appear after it.

For illustration, try searching for 'kf' in the MW dictionary. An empty
'vlex' element, rendered as a self-closing 'span', makes all later
siblings bold.

This is a hacky fix to a library that is just barely powerful enough to
do all that the site requires. When the site's requirements increase,
it will need to use a different library instead.