Add D3 visualization and some helper code
The visualization is a bubble chart of consonant frequencies in the
Bhagavad Gita. The site's JavaScript now contains some helper code for
creating bubble charts in the future. It also contains some helper code
for pie charts and bar graphs, both of which were considered and
eventually rejected for this particular visualization.