Modify lesson choices for 'Roots'
The main change is to discuss root suffixes earlier and add a lesson on
the neuter gender. These changes were made for a few reasons:

- The current pool of nouns is small and boring.
- Root suffixes are incredibly interesting and deserve to be discussed
  earlier in the unit.
- The use of new vocabulary throughout the lessons will help prevent
  the large number of verb forms from seeming like an onslaught.
- The lack of a neuter gender has limited the range of examples that
  can can draw from the Bhagavad Gita.
- The monomaniacal focus on verbs might turn some people off.

Another change is to discuss the gerund and the infinitive in the
same lesson. Both forms are verbal indeclinables that are easy to
form given prior lessons.