Change MW query form
Instead of using separate <select> fields for transliteration,
delegate to the site settings object. With this approach, forms are
less cluttered but still have the same functionality.

Of course, this approach assumes that the browser accepts cookies and
has JavaScript enabled, and it's possible that neither of these things
hold true.