Modify lesson choices in Roots
Specfiically, drop the lesson on causal verbs and add one on upasarga.

Roots is already a challenging chapter filled with new concepts:

- verb roots and classes
- verb prefixes
- passive verbs
- future and past tenses
- verbal adjectives and indeclinables

Layering causal verbs on top of this seems like too much. Aesthetically
it's tempting to bring the "journey to the center of the root" to a
close by analyzing the very nature of Sanskrit roots, but this unit
immediately follows the one that introduces what nouns are, so it's
perhaps too soon.

Moreover, the gain for a text like the Gita is small, since the text
only has about 80 causal verbs (as opposed to nearly 200 gerunds, only
3 of which are causative). Future verbs are rarer, but they introduce
some important concepts, including verb tense, ardhadhatuka, it,
multiple verb stems (but not in the sense of the athematic classes),
and classless stems. For that reason, the lesson on the near future
will remain as part of the chapter.

Instead of a lesson on causal verbs, the chapter will have a lesson
dedicated to the upasarga, as opposed to the tagalong nature that that
material has now. This puts some breathing room in the chapter and
creates some space to practice new concepts.

In that vein, the examples in this chapter should be much harder. Many
of them still restrict themselves to horses and heroes.