Add basic text system with admin panel
This is a basic system for querying texts and assigning URLs to
different text segment. This is a reimplementation of an older version
at nakula.learnsanskrit.org, and it aims to fix the many mistakes of
the original.

There's still plenty of work to do, but it's off to a good start.