Add various CSS to text and segment pages
For text pages: paged segments are shown Perseus-style in a horizontal
row of boxes. This is fairly usable but doesn't give enough feedback
to the user.

For segment pages: page links are simple arrows to the left and right.
As with the row of boxes described above, it's usable but doesn't give
enough feedback to the user.

For both, I considered giving feedback through tooltips, which worked
well to a point. But eventually it started to feel like a flimsy
solution to something that requires a bit more thought.