Fix incompatibility problems with dependencies
This is the first commit in quite some time, and a lot of things are
broken.

Currently: I can bring up the site, but the site goes 500 on a database
error. The database can't be built because fabric can't be loaded.
fabric can't be loaded because of context issues with Flask.

Revival plan:

1) Clean until the site can be loaded.
2) Clean until the site can be built with a single command (or two).
3) Add saner packaging for 'sanskrit' package
4) Automatic linting
5) Unit tests for all non-trivial pieces of code