Add interactive quiz system
This pulls the JSON data from the server and creates an interactive
quiz. The user is shown a question and enters in their own answer. When
a certain number of questions have been finished successfully (with a
small penalty for wrong answers), the quiz is "done."

Still TODO: better feedback if the user's answer is wrong.